# Move ptype2 shape handling to C