--- a/R/type-table.R
+++ b/R/type-table.R
@@ -34,7 +34,7 @@ vec_ptype2.table <- function(x, y, ..., x_arg = "", y_arg = "") {
 #' @method vec_ptype2.table table
 #' @export
 vec_ptype2.table.table <- function(x, y, ..., x_arg = "", y_arg = "") {
-  vec_shaped_ptype(new_table(), x, y)
+  vec_shaped_ptype(new_table(), x, y, x_arg = x_arg, y_arg = y_arg)
 }
 
 # ------------------------------------------------------------------------------