--- a/src/init.c
+++ b/src/init.c
@@ -108,6 +108,8 @@ extern SEXP vctrs_rep(SEXP, SEXP);
 extern SEXP vctrs_rep_each(SEXP, SEXP);
 extern SEXP vctrs_maybe_referenced_col(SEXP, SEXP);
 extern SEXP vctrs_new_df_unreferenced_col();
+extern SEXP vctrs_shaped_ptype(SEXP, SEXP, SEXP);
+extern SEXP vctrs_shape2(SEXP, SEXP);
 
 // Maturing
 // In the public header
@@ -236,6 +238,8 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_rep_each",                   (DL_FUNC) &vctrs_rep_each, 2},
   {"vctrs_maybe_referenced_col",       (DL_FUNC) &vctrs_maybe_referenced_col, 2},
   {"vctrs_new_df_unreferenced_col",    (DL_FUNC) &vctrs_new_df_unreferenced_col, 0},
+  {"vctrs_shaped_ptype",               (DL_FUNC) &vctrs_shaped_ptype, 3},
+  {"vctrs_shape2",                     (DL_FUNC) &vctrs_shape2, 2},
   {NULL, NULL, 0}
 };