--- a/src/names.c
+++ b/src/names.c
@@ -2,6 +2,7 @@
 #include "vctrs.h"
 #include "type-data-frame.h"
 #include "utils.h"
+#include "size.h"
 
 static void describe_repair(SEXP old_names, SEXP new_names);