--- a/src/shape.h
+++ b/src/shape.h
@@ -1,0 +1,8 @@
+#ifndef VCTRS_SHAPE_H
+#define VCTRS_SHAPE_H
+
+#include "vctrs.h"
+
+SEXP vec_shaped_ptype(SEXP ptype, SEXP x, SEXP y);
+
+#endif