--- a/src/shape.h
+++ b/src/shape.h
@@ -3,6 +3,8 @@
 
 #include "vctrs.h"
 
-SEXP vec_shaped_ptype(SEXP ptype, SEXP x, SEXP y);
+SEXP vec_shaped_ptype(SEXP ptype,
+                      SEXP x, SEXP y,
+                      struct vctrs_arg* p_x_arg, struct vctrs_arg* p_y_arg);
 
 #endif