--- a/src/size.c
+++ b/src/size.c
@@ -2,40 +2,18 @@
 #include "type-data-frame.h"
 #include "utils.h"
 #include "slice.h"
+#include "size.h"
 
 R_len_t rcrd_size(SEXP x);
 
-// [[ register(); include("vctrs.h") ]]
-SEXP vec_dim(SEXP x) {
-  SEXP dim = PROTECT(vec_bare_dim(x));
-
-  if (dim == R_NilValue) {
-    dim = r_int(Rf_length(x));
-  }
-
-  UNPROTECT(1);
-  return dim;
+// [[ register() ]]
+SEXP vctrs_dim(SEXP x) {
+  return vec_dim(x);
 }
 
-// [[ include("vctrs.h") ]]
-R_len_t vec_dim_n(SEXP x) {
-  return Rf_length(vec_dim(x));
-}
 // [[ register() ]]
 SEXP vctrs_dim_n(SEXP x) {
   return r_int(vec_dim_n(x));
-}
-
-// These versions return NULL and 0 for bare vectors. This is useful
-// to distinguish them from 1D arrays.
-
-// [[ include("vctrs.h") ]]
-SEXP vec_bare_dim(SEXP x) {
-  return Rf_getAttrib(x, R_DimSymbol);
-}
-// [[ include("vctrs.h") ]]
-R_len_t vec_bare_dim_n(SEXP x) {
-  return Rf_length(vec_bare_dim(x));
 }
 
 static inline R_len_t vec_raw_size(SEXP x) {