… src/slice-assign.c
… src/slice-chop.c
… src/slice.c
… src/strides.h
… src/subscript.c
… src/type2.c
… src/unspecified.c
… src/utils.h
… src/vctrs.h
… tests/testthat/error/test-shape.txt
… tests/testthat/test-shape.R
… tests/testthat/test-type-bare.R
… tests/testthat/test-type-table.R