This PR moves the common shape handling in `vec_ptype2()` down to C. This has the nice benefit of allowing us to mark the R level ptype2 methods for bare types with `stop_native_implementation()`.

- New `dim.c` and `dim.h` to hold dim specific functions, and make a few of the performance critical ones inline. This aligns better with `dim.R` too.

- New `shape.c` and `shape.h` to hold the common shape functionality

   - The idea of "shape" is tweaked slightly. It is now the dimensions vector with the first axis set to 0. It used to be all of the axes except the first, but this was a little harder to build on.

   - `vec_shape2()` is the main function here, it takes two sets of dimensions to find a common shape for and returns that common shape (as defined above). It replaces `shape_common()`.

   - `vec_shaped_ptype()` replaces `shape_match()` and is the entry point that `vec_ptype2()` uses. It computes the common shape from `x` and `y` and applies it to the `ptype`.

- When I added in the `stop_native_implementation()`, I learned that the complex methods were missing from `vec_ptype2_switch_native()`. I added them.

I detect essentially 0 loss of performance for non-shaped inputs.

```r
library(vctrs)

x <- 1
y <- 1

# Master
bench::mark(vec_ptype2(x, y), iterations = 1000000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, y)    673ns    866ns  1076072.        0B     107.

# This PR
bench::mark(vec_ptype2(x, y), iterations = 1000000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, y)    699ns    875ns  1071231.        0B     107.
```

Really nice performance now for array common types

```r
library(vctrs)

x <- array(1, dim = c(2, 1, 4, 6))
y <- array(1, dim = c(3, 5, 1, 6))

# Master
bench::mark(vec_ptype2(x, y), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, y)   19.9µs   23.4µs    39464.    99.1KB     54.5

# This PR
bench::mark(vec_ptype2(x, y), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_ptype2(x, y)    935ns   1.31µs   744330.        0B     89.3
```