@DavisVaughan 2020-04-08
@lionel- (and FYI for @hadley) we now throw `vctrs_error_incompatible_type` errors in `vec_ptype2()` when two shapes are incompatible. The error message is much better too.

``` r
library(vctrs)

x <- array(integer(), dim = c(0, 2, 3))
y <- array(integer(), dim = c(0, 1, 4))
```

```r
# Master
vec_ptype2(x, y)
#> Error: Incompatible lengths: 3, 4.

# This PR
vec_ptype2(x, y)
#> Error: Can't combine <integer[,2,3]> and <integer[,1,4]>.
#> x Incompatible sizes 3 and 4 along axis 3.
```

```r
# Master
vec_ptype2(x, y, x_arg = "foo", y_arg = "bar")
#> Error: Incompatible lengths: 3, 4.

# This PR
vec_ptype2(x, y, x_arg = "foo", y_arg = "bar")
#> Error: Can't combine `foo` <integer[,2,3]> and `bar` <integer[,1,4]>.
#> x Incompatible sizes 3 and 4 along axis 3.
```

@lionel- I rebased to bring in utilities from the type-error changes, but if you look at just the most recent commit here then that is where the changes are. Essentially I'm now threading all of the info required to give a good error message through `vec_shape2()`.