@hadley 2020-04-08
The error messages are a lot better, but I'm a little worried about using "axis" here; I think dimension would be clearer to most R users.