@DavisVaughan 2020-04-08
How strongly do you feel about that? I'm fairly confident that axis is the right technical term here. If we changed it, the error message would be different from the terminology used in the underlying code.

I like that axis specifies the direction along the array, and dimension specifies the measure along that axis (like the length). i.e. "the second axis has dimension 3". To me, that terminology also makes the code much easier to follow.