@hadley 2020-04-08
I don't feel too strongly about it, given that mostly people won't see these messages. But I think most people would describe a matrix as having two dimensions, not having two axes.